Replace text fields with dialog for user input
replaced permanent text field above graph as a means of getting user
input and replaced it with a button that invokes a dialog window. In
this window, the user may input polynomial functions as before, with the
added benefit of setting its color too. Some bugs are unresolved, like
the DialogPane stage being too big and having white space at the bottom.